test_client/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.3.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
jinja2>=3.0.0 